Refactored TransmissionFixture into a common base class, to share with VuzeFixture tests.
